06_gmail_classifier/requirements.txt
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
requests
beautifulsoup4
streamlit
pandas
python-dotenv
schedule
